Neatened all code
