+++ Stage.py
@@ -6,18 +6,15 @@ Last Modified: R. Nance 5/15/2018
 Originating Branch: Master
 Originally Created: R. Nance 12/2017
 '''
-import smbus
 from helper import *
 import time
 
 class Stage:
 
-    def __init__(self, address, position, bus):
+    def __init__(self, position):
         self.position = position
-        self.address = address
-        self.bus = smbus.SMBus(bus)
         self.home = 6000
-    bus = smbus.SMBus(1)
+
 
     def getPosition(self):
         return int(self.position)
@@ -67,20 +64,6 @@ class Stage:
 
         return command
 
-    def write(self, command):
-        bus = smbus.SMBus(1)
-        #bus.write_i2c_block_data(self.address, 0, command)
-        ##############CHANGED TO 1 BUT SHOULD BE ZERO
-
-        print(commandToString(command))  # print the command in  a user readable format.
-
-        bus.write_i2c_block_data(self.address, 0, command)
-
-    def write1(self, command):
-        bus = smbus.SMBus(1)
-        #bus.write_i2c_block_data(self.address, 0, command)
-        bus.write_i2c_block_data(0x32, 0, command)
-
     def sendCommand(self, commandCode, commandVars):
         commandToSend = self.buildCommand(commandCode, commandVars)
 
@@ -126,22 +109,29 @@ class Stage:
         P is position, hex representation of encoder counts
         E is error count. How far is the stage from where it is supposed to be?
         """
-        bus = self.bus
-        self.sendCommandNoVars('10')  #send query asking about motor status and position
-        temp = bus.read_i2c_block_data(0x32, 0)  #store incoming data from motor in list
+
+        self.sendCommandNoVars('10')  # send query asking about motor status and position
+        time.sleep(0.2)
+        temp = self.read()  # store incoming data from motor in list
 
         rcvEncodedPosition = ''
         for element in range(8):
-            rcvEncodedPosition += str(chr(temp[11+element]))
+            #print(temp[6])
+            #print(13 + element)
+            #print(temp[int(13 + element)])
+            #print(chr(temp[11 + element]))
+            rcvEncodedPosition += str(temp[13 + element])
+        #print(rcvEncodedPosition)
         position = int(rcvEncodedPosition, 16)
+        print('The current position Reported by M3LS is : ', position)
         return position
-
     def goToLocation(self, location):
         """
         Sends the stage to the location specified, in encoder counts
         :param location: a location in encoder counts
         :return: NA
         """
+        #print(encodeToCommand(location)) ###FOR DEBUGGING PURPOSES######
         self.sendCommand('08', encodeToCommand(location))
 
     def returnHome(self):
@@ -151,32 +141,6 @@ class Stage:
         """
         self.goToLocation(self.home)
 
-    def read(self):
-        """
-        Reads from the output register of the stage
-        :return: List of signed values that reprsent what is on the output register of the stage
-        """
-        bus = self.bus
-        temp = bus.read_i2c_block_data(self.address, 0)
-        print('temp', temp)
-        returnBuffer = []
-        for i in temp:
-            returnBuffer += str(chr(int(i)))
-
-        return returnBuffer
-
-    def zMove(self, direction, encoderCounts):
-
-        """
-
-
-        :param direction: The direction for Z to move. 1= up 0 = down
-        :param encoderCounts: number of encoder counts to move
-        :return: NA
-
-        """
-        command = '06 ' + str(direction)
-        self.sendCommand(command, encodeToCommand(encoderCounts))
